Added compose and post Tweet functionality
User can now compose and post a tweet in his own profile. Added several helping methods to complete the request, an added a delegate for the ComposeViewController, since it could be called from different parts of the app, and it now also returns data to it's sender (in this case the TimelineViewController). GIF with new tweet composing UX flow: http://g.recordit.co/Aauv5NAVnd.gif